# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Parking Management API",
    "description": "API for managing users, parking levels, vehicles, and transactions for a parking facility.\nThis documentation provides a base for all the resources defined in the database schema.\n",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "http://www.example.com/support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Production Server"
    }
  ],
...
  "components": {
    "schemas": {
      "BaseResponse": {
        "required": ["success", "message"],
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string", "example": "Success" }
        }
      },
      "User": {
        "type": "object",
        "required": ["id", "fullname", "username", "role"],
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "fullname": { "type": "string", "example": "John Doe" },
          "username": { "type": "string", "example": "john.doe" },
          "role": { "type": "string", "enum": ["admin", "member"], "example": "member" }
        }
      },
      "UserCreate": {
        "type": "object",
        "required": ["fullname", "username", "password", "role"],
        "properties": {
          "fullname": { "type": "string", "example": "Jane Doe" },
          "username": { "type": "string", "example": "jane.doe" },
          "password": { "type": "string", "format": "password", "example": "strongpassword123" },
          "role": { "type": "string", "enum": ["admin", "member"], "example": "admin" }
        }
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "fullname": { "type": "string", "example": "Jane D. Smith" },
          "username": { "type": "string", "example": "jane.smith" },
          "role": { "type": "string", "enum": ["admin", "member"], "example": "member" }
        }
      },
...
      "MembershipPlan": {
        "type": "object",
        "required": ["id", "cost", "period"],
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "cost": { "type": "number", "format": "double", "description": "Cost of the membership plan.", "example": 150.00 },
          "period": { "type": "integer", "description": "Duration of the plan in months.", "example": 12 }
        }
      },
      "MembershipPlanCreate": {
        "type": "object",
        "required": ["cost", "period"],
        "properties": {
          "cost": { "type": "number", "format": "double", "example": 50.00 },
          "period": { "type": "integer", "example": 6 }
        }
      },
...
      "Price": {
        "type": "object",
        "required": ["id", "categoryId", "amount", "type", "isActive"],
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "categoryId": { "type": "integer", "example": 1 },
          "amount": { "type": "number", "format": "double", "example": 5000 },
          "type": { "type": "string", "description": "Business rule type for the price.", "example": "INITIAL_BLOCK" },
          "blockHours": { "type": "integer", "nullable": true, "description": "Duration in hours for INITIAL_BLOCK type.", "example": 2 },
          "isActive": { "type": "boolean", "default": true },
          "validFrom": { "type": "string", "format": "date-time", "nullable": true },
          "validUntil": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "PriceCreate": {
        "type": "object",
        "required": ["categoryId", "amount", "type"],
        "properties": {
          "categoryId": { "type": "integer", "example": 2 },
          "amount": { "type": "number", "format": "double", "example": 10000 },
          "type": { "type": "string", "description": "Business rule type for the price.", "example": "INITIAL_BLOCK" },
          "blockHours": { "type": "integer", "nullable": true, "description": "Duration in hours for INITIAL_BLOCK type.", "example": 2 },
          "isActive": { "type": "boolean", "default": true },
          "validFrom": { "type": "string", "format": "date-time", "nullable": true },
          "validUntil": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
      "ApiErrorResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/BaseResponse" },
          {
            "type": "object",
            "required": ["data"],
            "properties": {
              "data": {
                "type": "object",
                "required": ["errorCode"],
                "properties": {
                  "errorCode": { "type": "string", "description": "A unique, machine-readable error code.", "example": "RESOURCE_NOT_FOUND" },
                  "errorMeta": {
                    "type": "array",
                    "description": "Optional metadata for validation errors.",
                    "items": {
                      "type": "object",
                      "properties": {
                        "field": { "type": "string", "example": "password" },
                        "issue": { "type": "string", "example": "Password must be at least 8 characters long." }
                      }
                    }
                  }
                }
              }
            }
          }
        ]
      },
      "UserResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/BaseResponse" },
          { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/User" } } }
        ]
      },
      "UserArrayResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/BaseResponse" },
          { "type": "object", "required": ["data"], "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/User" } } } }
        ]
      },
...
      "MembershipPlanResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/BaseResponse" },
          { "type": "object", "required": ["data"], "properties": { "data": { "$ref": "#/components/schemas/MembershipPlan" } } }
        ]
      },
      "MembershipPlanArrayResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/BaseResponse" },
          { "type": "object", "required": ["data"], "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/MembershipPlan" } } } }
        ]
      },
...
      "PriceResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/BaseResponse" },
          { "type": "object", "required": ["data"], "properties": { "data": { "$ref": "#/components/schemas/Price" } } }
        ]
      },
      "PriceArrayResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/BaseResponse" },
          { "type": "object", "required": ["data"], "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Price" } } } }
        ]
      },
...
      "SuccessResponse": {
        "$ref": "#/components/schemas/BaseResponse"
      }
    },
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List all users (Admin only)",
        "description": "Retrieves a list of all user accounts. This operation is restricted to admin users.",
        "operationId": "listUsers",
        "responses": {
          "200": { "description": "A list of users.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserArrayResponse" } } } },
          "403": { "description": "Forbidden.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user (Admin only)",
        "description": "Creates a new user account in the system. This operation is restricted to admin users.",
        "operationId": "createUser",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserCreate" } } } },
        "responses": {
          "201": { "description": "User created successfully", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } } },
          "400": { "description": "Bad Request.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
          "403": { "description": "Forbidden.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID (Admin only)",
        "description": "Retrieves the details of a specific user by their unique ID. This operation is restricted to admin users.",
        "operationId": "getUserById",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } } },
          "403": { "description": "Forbidden.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
          "404": { "description": "User not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update a user (Admin only)",
        "description": "Updates a user's details. This operation is restricted to admin users.",
        "operationId": "updateUser",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserUpdate" } } } },
        "responses": {
          "200": { "description": "User updated successfully", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } } },
          "400": { "description": "Bad Request.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
          "403": { "description": "Forbidden.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
          "404": { "description": "User not found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete a user (Admin only)",
        "description": "Deletes a user from the system. This operation is restricted to admin users.",
        "operationId": "deleteUser",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "User deleted successfully", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "403": { "description": "Forbidden.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
          "404": { "description": "User not found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } }
        }
      }
    },
...
    "/membership-plans": {
      "get": {
        "tags": ["Membership Plans"],
        "summary": "List all membership plans",
        "description": "Returns a list of all available membership plans.",
        "operationId": "listMembershipPlans",
        "security": [],
        "responses": {
          "200": { "description": "A list of membership plans.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MembershipPlanArrayResponse" } } } }
        }
      },
      "post": {
        "tags": ["Membership Plans"],
        "summary": "Create a membership plan (Admin only)",
        "description": "Creates a new membership plan. Restricted to admins.",
        "operationId": "createMembershipPlan",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MembershipPlanCreate" } } } },
        "responses": {
          "201": { "description": "Membership plan created successfully", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MembershipPlanResponse" } } } },
          "400": { "description": "Bad Request.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
          "403": { "description": "Forbidden.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } }
        }
      }
    },
    "/membership-plans/{id}": {
      "get": {
        "tags": ["Membership Plans"],
        "summary": "Get membership plan by ID",
        "description": "Retrieves details for a specific membership plan.",
        "operationId": "getMembershipPlanById",
        "security": [],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MembershipPlanResponse" } } } },
          "404": { "description": "Membership plan not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } }
        }
      },
      "put": {
        "tags": ["Membership Plans"],
        "summary": "Update a membership plan (Admin only)",
        "description": "Updates a membership plan. Restricted to admins.",
        "operationId": "updateMembershipPlan",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MembershipPlanCreate" } } } },
        "responses": {
          "200": { "description": "Membership plan updated successfully", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MembershipPlanResponse" } } } },
          "400": { "description": "Bad Request.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
          "403": { "description": "Forbidden.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
          "404": { "description": "Membership plan not found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Membership Plans"],
        "summary": "Delete a membership plan (Admin only)",
        "description": "Deletes a membership plan. Restricted to admins.",
        "operationId": "deleteMembershipPlan",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Membership plan deleted successfully", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "403": { "description": "Forbidden.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
          "404": { "description": "Membership plan not found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } }
        }
      }
    },
...
    "/prices": {
      "get": {
        "tags": ["Prices"],
        "summary": "List all prices",
        "description": "Returns a list of all price configurations.",
        "operationId": "listPrices",
        "security": [],
        "responses": {
          "200": { "description": "A list of prices.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PriceArrayResponse" } } } }
        }
      },
      "post": {
        "tags": ["Prices"],
        "summary": "Create a price (Admin only)",
        "description": "Creates a new price configuration. Restricted to admins.",
        "operationId": "createPrice",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PriceCreate" } } } },
        "responses": {
          "201": { "description": "Price created successfully", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PriceResponse" } } } },
          "400": { "description": "Bad Request.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
          "403": { "description": "Forbidden.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } }
        }
      }
    },
    "/prices/{id}": {
      "get": {
        "tags": ["Prices"],
        "summary": "Get price by ID",
        "description": "Retrieves details for a specific price configuration.",
        "operationId": "getPriceById",
        "security": [],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PriceResponse" } } } },
          "404": { "description": "Price not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } }
        }
      },
      "put": {
        "tags": ["Prices"],
        "summary": "Update a price (Admin only)",
        "description": "Updates a price configuration. Restricted to admins.",
        "operationId": "updatePrice",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PriceCreate" } } } },
        "responses": {
          "200": { "description": "Price updated successfully", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PriceResponse" } } } },
          "400": { "description": "Bad Request.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
          "403": { "description": "Forbidden.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
          "404": { "description": "Price not found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Prices"],
        "summary": "Delete a price (Admin only)",
        "description": "Deletes a price configuration. Restricted to admins.",
        "operationId": "deletePrice",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Price deleted successfully", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "403": { "description": "Forbidden.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } },
          "404": { "description": "Price not found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } } }
        }
      }
    },
...
  }
}